Testing data for BLEU Score
Modified how data is tokenized to allow for all testing data to be added to word_index.
Testing data is fed to model and response is output to separate file to be sent to website to determine score.
All dialogue is required to be in one file so added to pre-processing.